fixing datatype of lat/lon

+++ snap-netcdf/src/main/java/org/esa/snap/dataio/netcdf/metadata/profiles/cf/CfGeocodingPart.java
@@ -179,14 +179,14 @@ public class CfGeocodingPart extends ProfilePartIO {
     }
 
     private void addGeographicCoordinateVariables(NFileWriteable ncFile, GeoPos ul, GeoPos br) throws IOException {
-        final NVariable lat = ncFile.addVariable("lat", DataType.FLOAT, null, "lat");
+        final NVariable lat = ncFile.addVariable("lat", DataType.DOUBLE, null, "lat");
         lat.addAttribute("units", "degrees_north");
         lat.addAttribute("long_name", "latitude");
         lat.addAttribute("standard_name", "latitude");
         lat.addAttribute(Constants.VALID_MIN_ATT_NAME, br.getLat());
         lat.addAttribute(Constants.VALID_MAX_ATT_NAME, ul.getLat());
 
-        final NVariable lon = ncFile.addVariable("lon", DataType.FLOAT, null, "lon");
+        final NVariable lon = ncFile.addVariable("lon", DataType.DOUBLE, null, "lon");
         lon.addAttribute("units", "degrees_east");
         lon.addAttribute("long_name", "longitude");
         lon.addAttribute("standard_name", "longitude");
@@ -195,12 +195,12 @@ public class CfGeocodingPart extends ProfilePartIO {
     }
 
     private void addLatLonBands(final NFileWriteable ncFile, Dimension tileSize) throws IOException {
-        final NVariable lat = ncFile.addVariable("lat", DataType.FLOAT, tileSize, "y x");
+        final NVariable lat = ncFile.addVariable("lat", DataType.DOUBLE, tileSize, "y x");
         lat.addAttribute("units", "degrees_north");
         lat.addAttribute("long_name", "latitude coordinate");
         lat.addAttribute("standard_name", "latitude");
 
-        final NVariable lon = ncFile.addVariable("lon", DataType.FLOAT, tileSize, "y x");
+        final NVariable lon = ncFile.addVariable("lon", DataType.DOUBLE, tileSize, "y x");
         lon.addAttribute("units", "degrees_east");
         lon.addAttribute("long_name", "longitude coordinate");
         lon.addAttribute("standard_name", "longitude");
